% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{find_card}
\alias{find_card}
\title{Find a FIST card}
\usage{
find_card(colour, animal, number, cards = fist_cards())
}
\arguments{
\item{colour, animal, number}{Card features.}

\item{cards}{A card metadata data frame, such as `fist_cards()`.}
}
\value{
A one-row data frame.
}
\description{
Finds one card in a FIST card deck by its features.
}
\examples{
find_card("blue", "frog", 1)
}
