% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_fist_game}
\alias{run_fist_game}
\title{Run the FIST card game}
\usage{
run_fist_game(...)
}
\arguments{
\item{...}{Additional arguments passed to `shiny::runApp()`.}
}
\value{
The result of `shiny::runApp()`.
}
\description{
Starts the Shiny version of the FIST card task.
}
\examples{
\dontrun{
run_fist_game()
}
}
